Added line_items export to Excel; narrowed relationships criteria.
Created Contracts:lineitems action to export line items to Excel.  Also,
created show.xls.builder for contracts, (which currently copies the behavior
of lineitems.xls) to be used for quoting in the future.

The 'Replaced By' and 'Replaces' list in the Contracts:Edit view are now
narrowed down to contracts newer and older respectively than the currently
edited contract.