Overhaul of the permissions system; Cleanup of the before_filters
Added Role model which links to User via Permission.  This allows for more
fine-grained control of permissions in the application.  the app still uses
the old permissions system, but the goal is to convert over in short order.

Centralized the login_required before_filter to the application_controller.
To not require login, one must use skip_before_filter as in the graphs_controller.

Removed the @current_tab variable that was still being set for several controllers.
This was from the old menu layout, and is not needed, thus was removed.

Cleaned up the authorization_required? methods across the application.