Added contact information to Contract and 'note' field to LineItems
The edit/new form and the show view now match as far as field order and position.

These fields were added to enable preserving the data from the quote tool, when
using the renew contract function in the quoter tool, and when importing.

Corresponding changes to the YAMLExporter and the SupportQuoterTool will be made.