Drag-and-Drop for Line Items implemented.  Permissions changes.
Added Drag-and-Drop to re-order line items.  To make this effective, also
added permissions for anyone with the :contract_editor role to re-order, and do
limited mass updates, as well as individual edits and deletes.

Added a simple_show view that restricts the more confidential information of
customers.  This is now the default view for anyone who is not a contract_admin,
manager, or a member of one of the support_office/sales_office teams to see.

This will provide a way for call screeners to view contracts and still maintain
security.