Migrations changed; support_type validation removed; 2 features passing tests.
Modified some of the legacy migrations in order to be able to create a test
database from scratch.  There were some inconsistencies that had cropped up
between the current state of the database and the result of running the
migrations.  The current DB schema will now be recreated correctly when the
migrations are run.

the support level validations have been removed, since they were not passing
the test - likely any future levels that would have been added would not take
effect until after a server reboot.  They have been removed as a workaround.

contract_management and authenticate features are passing.